final/requirements.txt
streamlit==1.31.0
google-generativeai==0.3.2
passlib==1.7.4
plotly==5.19.0
python-dotenv==1.0.0
streamlit-card==1.0.2
streamlit-elements==0.1.0
streamlit-extras==0.3.6
pandas==2.2.1
psycopg2-binary==2.9.9
pyarrow==14.0.2
streamlit-js-eval==0.1.5
